requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
pypdf
langchain-qdrant
qdrant-client
fastapi
uvicorn
python-dotenv
litellm
langdetect
langchain-openai
pandas
requests
openpyxl
unstructured
tabulate
networkx
openpyxl
